refactor: rename render_state.rs and runtime_state.rs
